% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrSVmain}
\alias{wrSVmain}
\title{Write code for main block of server.R}
\usage{
wrSVmain(prefix, grouping.var = "group", subst = "")
}
\arguments{
\item{prefix}{file prefix}
}
\description{
Write code for main block of server.R
}
